Refactor common place importer methods to base importer
